+++ infra/modules/azure_cosmos_account/tests/setup/providers.tf
@@ -12,14 +12,18 @@ terraform {
       source  = "hashicorp/random"
       version = "~> 3.7"
     }
+    time = {
+      source  = "hashicorp/time"
+      version = "~> 0.12"
+    }
   }
 }
 
 provider "azurerm" {
   features {
     key_vault {
-      purge_soft_delete_on_destroy    = true
-      recover_soft_deleted_key_vaults = false
+      purge_soft_delete_on_destroy    = false
+      recover_soft_deleted_key_vaults = true
     }
     resource_group {
       prevent_deletion_if_contains_resources = false
@@ -28,4 +32,7 @@ provider "azurerm" {
 }
 
 provider "random" {
+}
+
+provider "time" {
 }