+++ infra/modules/azure_function_app/function_app_slot.tf
@@ -93,6 +93,7 @@ resource "azurerm_linux_function_app_slot" "this" {
 
   lifecycle {
     ignore_changes = [
+      site_config[0].health_check_eviction_time_in_min,
       app_settings["WEBSITE_HEALTHCHECK_MAXPINGFAILURES"],
       tags["hidden-link: /app-insights-conn-string"],
       tags["hidden-link: /app-insights-instrumentation-key"],